Fixing bugs and improving code quality
When opening Add or Edit component, instead of creating a new
connection, and pulling again data from Firebase, its only pulls once in
the App, and passes the data to the Edit and Add component.

Replace a few variables with a single book object, this contains the
necessary data for App component. This fixes the bugs where books appear
under wrong years, and have incorrect book covers.

Rename a few variables and functions in all components for better
readability.